FIP Layer: Create a base class for FIP image layers
