+++ test/test.vcxproj.filters
@@ -138,5 +138,8 @@
     <ClCompile Include="test_ageing_counter.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\FIPLayer.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
